+++ src/shortcuts.js
@@ -74,12 +74,8 @@ export function initializeShortcuts() {
             await executeBatchAddTasks();
         });
 
-        ipcRenderer.on('menu-edit-task', async () => {
-            await executeEditTask();
-        });
-
-        ipcRenderer.on('menu-batch-edit-tasks', async () => {
-            await executeBatchEditTasks();
+        ipcRenderer.on('menu-edit-tasks', async () => {
+            await executeEditTasks();
         });
 
         ipcRenderer.on('menu-remove-tasks', async () => {
@@ -133,6 +129,16 @@ export function initializeShortcuts() {
 
         Mousetrap.bind(['command+alt+t', 'ctrl+shift+t'], async () => {
             await executeAddTask();
+            return false;
+        });
+
+        Mousetrap.bind(['command+alt+b', 'ctrl+shift+b'], async () => {
+            await executeBatchAddTasks();
+            return false;
+        });
+
+        Mousetrap.bind(['command+alt+e', 'ctrl+shift+e'], async () => {
+            await executeEditTasks();
             return false;
         });
     }
@@ -191,29 +197,19 @@ async function executeBatchAddTasks() {
     await store.dispatch(setBatchAddTasksManagerOptions({ visible: true }));
 }
 
-async function executeEditTask() {
+async function executeEditTasks() {
     const selectedTaskIds = getSelectedTaskIds(store.getState());
 
-    if (selectedTaskIds.length !== 1) {
-        message.error('Please select a task');
-        return;
+    if (selectedTaskIds.length === 1) {
+        await store.dispatch(setTaskEditionManagerOptions({
+            visible: true,
+            taskId: selectedTaskIds[0]
+        }));
+    } else if (selectedTaskIds.length > 1 && selectedTaskIds.length <= 50) {
+        await store.dispatch(setBatchEditTasksManagerOptions({ visible: true }));
+    } else {
+        message.error('Please select one or more tasks (maximum 50)');
     }
-
-    await store.dispatch(setTaskEditionManagerOptions({
-        visible: true,
-        taskId: selectedTaskIds[0]
-    }));
-}
-
-async function executeBatchEditTasks() {
-    const selectedTaskIds = getSelectedTaskIds(store.getState());
-
-    if (selectedTaskIds.length <= 1 || selectedTaskIds.length > 50) {
-        message.error('Please select two or more tasks (maximum 50)');
-        return;
-    }
-
-    await store.dispatch(setBatchEditTasksManagerOptions({ visible: true }));
 }
 
 async function executeRemoveTasks() {